+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ## :zap: New Features
 
-* Nothing new :disappointed:
+- Added `images` command
 
 ## :lady_beetle: Fixes
 
-* No fixes for now :sweat_smile:
+- No fixes for now :sweat_smile: